+++ relay-event-normalization/src/normalize/span/tag_extraction.rs
@@ -167,10 +167,13 @@ struct SharedTags {
     transaction_method: Annotated<String>,
     transaction_op: Annotated<String>,
     transaction: Annotated<String>,
+    user_city: Annotated<String>,
     user_country_code: Annotated<String>,
     user_email: Annotated<String>,
     user_id: Annotated<String>,
     user_ip: Annotated<String>,
+    user_region: Annotated<String>,
+    user_subdivision: Annotated<String>,
     user_subregion: Annotated<String>,
     user_username: Annotated<String>,
     user: Annotated<String>,
@@ -195,10 +198,13 @@ impl SharedTags {
             transaction_method,
             transaction_op,
             transaction,
+            user_city,
             user_country_code,
             user_email,
             user_id,
             user_ip,
+            user_region,
+            user_subdivision,
             user_subregion,
             user_username,
             user,
@@ -251,6 +257,9 @@ impl SharedTags {
         if tags.transaction.value().is_none() {
             tags.transaction = transaction.clone();
         };
+        if tags.user_city.value().is_none() {
+            tags.user_city = user_city.clone();
+        }
         if tags.user_country_code.value().is_none() {
             tags.user_country_code = user_country_code.clone();
         };
@@ -263,6 +272,12 @@ impl SharedTags {
         if tags.user_ip.value().is_none() {
             tags.user_ip = user_ip.clone();
         };
+        if tags.user_region.value().is_none() {
+            tags.user_region = user_region.clone();
+        };
+        if tags.user_subdivision.value().is_none() {
+            tags.user_subdivision = user_subdivision.clone();
+        }
         if tags.user_subregion.value().is_none() {
             tags.user_subregion = user_subregion.clone();
         };
@@ -300,18 +315,25 @@ fn extract_shared_tags(event: &Event) -> SharedTags {
             tags.user_email = user_email.clone().into();
         }
 
-        // We only want this on frontend or mobile modules.
-        let should_extract_geo = (event.context::<BrowserContext>().is_some()
-            && event.platform.as_str() == Some("javascript"))
-            || MOBILE_SDKS.contains(&event.sdk_name());
+        if let Some(geo) = user.geo.value() {
+            if let Some(city) = geo.city.value() {
+                tags.user_city = city.clone().into();
+            }
 
-        if should_extract_geo
-            && let Some(country_code) = user.geo.value().and_then(|geo| geo.country_code.value())
-        {
-            tags.user_country_code = country_code.to_owned().into();
-            if let Some(subregion) = Subregion::from_iso2(country_code.as_str()) {
-                let numerical_subregion = subregion as u8;
-                tags.user_subregion = numerical_subregion.to_string().into();
+            if let Some(country_code) = geo.country_code.value() {
+                tags.user_country_code = country_code.clone().into();
+                if let Some(subregion) = Subregion::from_iso2(country_code.as_str()) {
+                    let numerical_subregion = subregion as u8;
+                    tags.user_subregion = numerical_subregion.to_string().into();
+                }
+            }
+
+            if let Some(region) = geo.region.value() {
+                tags.user_region = region.clone().into();
+            }
+
+            if let Some(subdivision) = geo.subdivision.value() {
+                tags.user_subdivision = subdivision.clone().into();
             }
         }
     }
@@ -3052,7 +3074,10 @@ LIMIT 1
                     "email": "admin@sentry.io",
                     "username": "admin",
                     "geo": {
-                        "country_code": "US"
+                        "country_code": "AT",
+                        "city": "Vienna",
+                        "subdivision": "Vienna",
+                        "region": "Austria"
                     }
                 },
                 "spans": [
@@ -3085,65 +3110,11 @@ LIMIT 1
         assert_eq!(get_value!(span.sentry_tags.user_ip!), "127.0.0.1");
         assert_eq!(get_value!(span.sentry_tags.user_username!), "admin");
         assert_eq!(get_value!(span.sentry_tags.user_email!), "admin@sentry.io");
-        assert_eq!(get_value!(span.sentry_tags.user_country_code!), "US");
-        assert_eq!(get_value!(span.sentry_tags.user_subregion!), "21");
-    }
-
-    #[test]
-    fn not_extract_geo_location_if_not_browser() {
-        let json = r#"
-            {
-                "type": "transaction",
-                "platform": "python",
-                "start_timestamp": "2021-04-26T07:59:01+0100",
-                "timestamp": "2021-04-26T08:00:00+0100",
-                "transaction": "foo",
-                "contexts": {
-                    "trace": {
-                        "trace_id": "ff62a8b040f340bda5d830223def1d81",
-                        "span_id": "bd429c44b67a3eb4"
-                    },
-                    "browser": {
-                        "name": "Chrome"
-                    }
-                },
-                "user": {
-                    "id": "1",
-                    "email": "admin@sentry.io",
-                    "username": "admin",
-                    "geo": {
-                        "country_code": "US"
-                    }
-                },
-                "spans": [
-                    {
-                        "op": "http.client",
-                        "span_id": "bd429c44b67a3eb1",
-                        "start_timestamp": 1597976300.0000000,
-                        "timestamp": 1597976302.0000000,
-                        "trace_id": "ff62a8b040f340bda5d830223def1d81"
-                    }
-                ]
-            }
-        "#;
-
-        let mut event = Annotated::<Event>::from_json(json).unwrap();
-
-        normalize_event(
-            &mut event,
-            &NormalizationConfig {
-                enrich_spans: true,
-                ..Default::default()
-            },
-        );
-
-        let spans = get_value!(event.spans!);
-        let span = &spans[0];
-
-        let tags = span.value().unwrap().sentry_tags.value().unwrap();
-
-        assert_eq!(tags.get_value("user.geo.subregion"), None);
-        assert_eq!(tags.get_value("user.geo.country_code"), None);
+        assert_eq!(get_value!(span.sentry_tags.user_country_code!), "AT");
+        assert_eq!(get_value!(span.sentry_tags.user_city!), "Vienna");
+        assert_eq!(get_value!(span.sentry_tags.user_region!), "Austria");
+        assert_eq!(get_value!(span.sentry_tags.user_subdivision!), "Vienna");
+        assert_eq!(get_value!(span.sentry_tags.user_subregion!), "155");
     }
 
     #[test]